# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Milchvieh Datenanalyse-Assistent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /datasets/{datasetId}/files:
    get:
      tags: [Files]
      operationId: listFiles
      summary: List files in a dataset
      parameters:
        - $ref: "#/components/parameters/DatasetId"
      responses:
        "200":
          description: Files.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceFile"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Files]
      operationId: registerFile
      summary: Register an uploaded file and start ingestion
      parameters:
        - $ref: "#/components/parameters/DatasetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceFileInput"
      responses:
        "201":
          description: Registered; ingestion started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceFile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /files/{fileId}:
    get:
      tags: [Files]
      operationId: getFile
      summary: Get a file with ingestion status and column mapping
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: File.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceFileDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Files]
      operationId: deleteFile
      summary: Delete a file and its parsed data
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"
...
  /datasets/{datasetId}/reports:
    get:
      tags: [Reports]
      operationId: listReports
      summary: List reports for a dataset
      parameters:
        - $ref: "#/components/parameters/DatasetId"
      responses:
        "200":
          description: Reports.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Reports]
      operationId: generateReport
      summary: Generate a report now
      parameters:
        - $ref: "#/components/parameters/DatasetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /reports/{reportId}:
    get:
      tags: [Reports]
      operationId: getReport
      summary: Get a report
      parameters:
        - $ref: "#/components/parameters/ReportId"
      responses:
        "200":
          description: Report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          $ref: "#/components/responses/NotFound"
  /masterdata:
    get:
      tags: [MasterData]
      operationId: listMasterData
      summary: List operator-maintained reference data
      responses:
        "200":
          description: Master data entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MasterDataEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [MasterData]
      operationId: createMasterData
      summary: Create a reference data entry (operator only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasterDataInput"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MasterDataEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /masterdata/{entryId}:
    patch:
      tags: [MasterData]
      operationId: updateMasterData
      summary: Update a reference data entry (operator only)
      parameters:
        - $ref: "#/components/parameters/EntryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasterDataUpdate"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MasterDataEntry"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [MasterData]
      operationId: deleteMasterData
      summary: Delete a reference data entry (operator only)
      parameters:
        - $ref: "#/components/parameters/EntryId"
      responses:
        "204":
          description: Deleted.
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    DatasetId:
      name: datasetId
      in: path
      required: true
      schema:
        type: string
    FileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
...
    ReportId:
      name: reportId
      in: path
      required: true
      schema:
        type: string
    EntryId:
      name: entryId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Authentication required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Not allowed for this role.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SourceFile:
      type: object
      required: [id, datasetId, name, status, createdAt]
      properties:
        id:
          type: string
        datasetId:
          type: string
        name:
          type: string
        contentType:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        status:
          type: string
          enum: [uploaded, parsing, mapping, ready, error]
        kind:
          type: ["string", "null"]
          enum: [excel, csv, herd_export, pdf, ppt, other, null]
        rowCount:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SourceFileInput:
      type: object
      required: [name, objectPath]
      properties:
        name:
          type: string
          minLength: 1
        objectPath:
          type: string
          minLength: 1
        contentType:
          type: string
        size:
          type: integer
          minimum: 0
    DetectedColumn:
      type: object
      required: [sourceColumn]
      properties:
        sourceColumn:
          type: string
        sampleValues:
          type: array
          items:
            type: string
        canonicalField:
          type: ["string", "null"]
        confidence:
          type: ["number", "null"]
    SourceFileDetail:
      type: object
      required: [id, datasetId, name, status, createdAt, columns]
      properties:
        id:
          type: string
        datasetId:
          type: string
        name:
          type: string
        contentType:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        status:
          type: string
          enum: [uploaded, parsing, mapping, ready, error]
        kind:
          type: ["string", "null"]
          enum: [excel, csv, herd_export, pdf, ppt, other, null]
        rowCount:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        columns:
          type: array
          items:
            $ref: "#/components/schemas/DetectedColumn"
        previewRows:
          type: array
          items:
            type: object
            additionalProperties: true
...
    ChartSeries:
      type: object
      required: [key, label]
      properties:
        key:
          type: string
        label:
          type: string
    Chart:
      type: object
      required: [id, type, title, data]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [line, bar, area, pie, scatter, table]
        title:
          type: string
        description:
          type: ["string", "null"]
        xKey:
          type: ["string", "null"]
        series:
          type: array
          items:
            $ref: "#/components/schemas/ChartSeries"
        data:
          type: array
          items:
            type: object
            additionalProperties: true
        unit:
          type: ["string", "null"]
        basis:
          type: ["string", "null"]
...
    Report:
      type: object
      required: [id, datasetId, title, period, status, createdAt]
      properties:
        id:
          type: string
        datasetId:
          type: string
        title:
          type: string
        period:
          type: string
          enum: [weekly, monthly, quarterly, custom]
        summary:
          type: ["string", "null"]
        sections:
          type: array
          items:
            $ref: "#/components/schemas/ReportSection"
        status:
          type: string
          enum: [generating, ready, error]
        createdAt:
          type: string
          format: date-time
    ReportSection:
      type: object
      required: [title]
      properties:
        title:
          type: string
        content:
          type: ["string", "null"]
        charts:
          type: array
          items:
            $ref: "#/components/schemas/Chart"
    ReportInput:
      type: object
      required: [period]
      properties:
        period:
          type: string
          enum: [weekly, monthly, quarterly, custom]
        title:
          type: string
    MasterDataEntry:
      type: object
      required: [id, category, key, value, createdAt]
      properties:
        id:
          type: string
        category:
          type: string
        key:
          type: string
        value:
          type: string
        unit:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    MasterDataInput:
      type: object
      required: [category, key, value]
      properties:
        category:
          type: string
          minLength: 1
        key:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
        unit:
          type: string
        notes:
          type: string
    MasterDataUpdate:
      type: object
      properties:
        category:
          type: string
          minLength: 1
        key:
          type: string
          minLength: 1
        value:
          type: string
          minLength: 1
        unit:
          type: string
        notes:
          type: string
...